+++ tests/edges/test_bsplinecurve.py
@@ -152,7 +152,7 @@ class TestBSplineCurve2D(unittest.TestCase):
         bspline_curve2d = bspline_curves.bspline_curve2d_1
         point = volmdlr.Point2D(-0.31240117104573617, -2.8555856978321796)
 
-        bspline = vme.BSplineCurve2D.load_from_file(os.path.join(folder, "bg_bspline5_.json"))
+        bspline = vme.BSplineCurve2D.from_json(os.path.join(folder, "bg_bspline5_.json"))
         point1 = bspline.points[25]
         point2 = bspline.points[75]
 
@@ -180,7 +180,7 @@ class TestBSplineCurve2D(unittest.TestCase):
         self.assertAlmostEqual(bspline_curve2d.abscissa(point), 7.747599410268476)
 
     def test_line_intersections(self):
-        bspline_curve2d = DessiaObject.load_from_file(os.path.join(folder, "bsplinecurve2d_1.json"))
+        bspline_curve2d = DessiaObject.from_json(os.path.join(folder, "bsplinecurve2d_1.json"))
         line = curves.Line2D(volmdlr.Point2D(1.263163105753452, -0.002645572020392778),
                           volmdlr.Point2D(1.263163105753452, -0.001820963841291406))
 
@@ -312,13 +312,13 @@ class TestBSplineCurve2D(unittest.TestCase):
             self.assertTrue(point1.is_close(point2))
 
     def test_simplify(self):
-        bsplinecurve = vme.BSplineCurve3D.load_from_file(os.path.join(folder, "bsplinecurve_fullarc.json"))
+        bsplinecurve = vme.BSplineCurve3D.from_json(os.path.join(folder, "bsplinecurve_fullarc.json"))
         fullarc = bsplinecurve.simplify
         self.assertTrue(isinstance(fullarc, vme.FullArc3D))
 
     def test_direction_independent_is_close(self):
-        bsplinecurve1 = vme.BSplineCurve3D.load_from_file(os.path.join(folder, "bspline_curve1.json"))
-        bsplinecurve2 = vme.BSplineCurve3D.load_from_file(os.path.join(folder, "bspline_curve2.json"))
+        bsplinecurve1 = vme.BSplineCurve3D.from_json(os.path.join(folder, "bspline_curve1.json"))
+        bsplinecurve2 = vme.BSplineCurve3D.from_json(os.path.join(folder, "bspline_curve2.json"))
         self.assertTrue(bsplinecurve1.direction_independent_is_close(bsplinecurve2))
 
     def test_split_curve(self):
@@ -330,7 +330,7 @@ class TestBSplineCurve2D(unittest.TestCase):
         self.assertTrue(splitted_curves[1].end.is_close(volmdlr.Point2D(50.0, 5.0)))
 
         split_point = volmdlr.Point2D(0.04820589473987067, 0.011936395549382077)
-        bsplinecurve = vme.BSplineCurve2D.load_from_file(os.path.join(folder, "bsplinecurve_split_bug.json"))
+        bsplinecurve = vme.BSplineCurve2D.from_json(os.path.join(folder, "bsplinecurve_split_bug.json"))
         splitted_curves = bsplinecurve.split(split_point)
         self.assertTrue(splitted_curves[0].start.is_close(volmdlr.Point2D(0.04873977000999985, 0.011815456390639745)))
         self.assertTrue(splitted_curves[0].end.is_close(split_point))