+++ tests/surfaces/test_bsplinesurface3d.py
@@ -172,7 +172,7 @@ class TestBSplineSurface3D(unittest.TestCase):
                     for c, e in zip(computed[idx], expected[idx]):
                         self.assertAlmostEqual(c, e, delta=DELTA)
 
-        surface = surfaces.BSplineSurface3D.load_from_file(
+        surface = surfaces.BSplineSurface3D.from_json(
             os.path.join(folder, "bsplinesurface_derivatives_v_degree_1.json"))
         test_data = [((0.0, 0.41570203515189436), 2,
                      [[(2.686370456553301, -0.6157625276711683, 0.5584759391609816),
@@ -308,7 +308,7 @@ class TestBSplineSurface3D(unittest.TestCase):
         self.assertAlmostEqual(contour2d_dim.length(), 16.823814079415172, places=2)
 
     def test_periodicity(self):
-        bspline_suface = surfaces.BSplineSurface3D.load_from_file(os.path.join(folder, 'surface3d_8.json'))
+        bspline_suface = surfaces.BSplineSurface3D.from_json(os.path.join(folder, 'surface3d_8.json'))
         self.assertAlmostEqual(bspline_suface.x_periodicity,  0.8888888888888888)
         self.assertFalse(bspline_suface.y_periodicity)
 
@@ -321,7 +321,7 @@ class TestBSplineSurface3D(unittest.TestCase):
         self.assertAlmostEqual(volume, 3.97787, 2)
 
     def test_arc3d_to_2d(self):
-        bspline_surface = surfaces.BSplineSurface3D.load_from_file(
+        bspline_surface = surfaces.BSplineSurface3D.from_json(
             os.path.join(folder, 'BSplineSurface3D_with_Arc3D.json'))
         arc = vme.Arc3D.from_3_points(volmdlr.Point3D(-0.01, -0.013722146986970815, 0.026677756316261864),
                         volmdlr.Point3D(-0.01, 0.013517082603, 0.026782241839),
@@ -338,9 +338,9 @@ class TestBSplineSurface3D(unittest.TestCase):
         self.assertTrue(inv_prof.end.is_close(arc.end))
 
         # Strange case from step file
-        bspline_surface = surfaces.BSplineSurface3D.load_from_file(
+        bspline_surface = surfaces.BSplineSurface3D.from_json(
             os.path.join(folder, 'bsplinesurface_arc3d_to_2d_surface.json'))
-        arc = vme.Arc3D.load_from_file(os.path.join(folder, "bsplinesurface_arc3d_to_2d_arc3d.json"))
+        arc = vme.Arc3D.from_json(os.path.join(folder, "bsplinesurface_arc3d_to_2d_arc3d.json"))
         brep = bspline_surface.arc3d_to_2d(arc)[0]
         self.assertTrue(brep.start.is_close(volmdlr.Point2D(1, 0)))
 
@@ -363,7 +363,7 @@ class TestBSplineSurface3D(unittest.TestCase):
         self.assertAlmostEqual(original_length, length_after_transformation, places=6)
         self.assertTrue(point.is_close(point_test, 1e-6))
 
-        surface = surfaces.BSplineSurface3D.load_from_file(
+        surface = surfaces.BSplineSurface3D.from_json(
             os.path.join(folder, "bsplinesurface_smallbsplinecurve.json"))
         bsplinecurve3d = vme.BSplineCurve3D.load_from_file(
             os.path.join(folder, "bsplinesurface_smallbsplinecurve_curve.json"))